errors fixed
malformed Jinja template in backend/context/prompts/research/synthesize.md
missing graph registration in backend/graphs/init.py
fallback analytics rules in backend/nodes/generate_sql.py
SQL validation ordering in backend/nodes/validate_sql.py
DuckDB compatibility in backend/clients/duckdb_client.py
Decimal precision in backend/schemas.py
route-level analytics execution in backend/routes/chat.py

+++ multi-agent-backend/backend/schemas.py
@@ -11,6 +11,8 @@ import uuid
 from datetime import datetime
 from typing import Any, Literal
 
+from decimal import Decimal
+
 from pydantic import BaseModel, ConfigDict, Field
 
 ChatRole = Literal["system", "user", "assistant", "tool"]
@@ -121,7 +123,7 @@ class DecisionLogRead(BaseModel):
     output: dict[str, Any] | None = None
     evaluation_score: float | None = None
     latency_ms: int | None = None
-    cost_usd: float | None = None
+    cost_usd: Decimal | None = None
     prompt_tokens: int | None = None
     completion_tokens: int | None = None
     error: str | None = None
